+++ web_src/src/ui/BuildingBlocksSidebar/index.spec.tsx
@@ -1,6 +1,5 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
-import { DEFAULT_AGENT_CONTEXT } from "./agentChat";
 import { BuildingBlocksSidebar } from "./index";
 
 const defaultProps = {
@@ -8,7 +7,6 @@ const defaultProps = {
   onToggle: vi.fn(),
   blocks: [],
   canvasZoom: 1,
-  agentContext: DEFAULT_AGENT_CONTEXT,
 };
 
 describe("BuildingBlocksSidebar", () => {
