+++ web_src/src/ui/BuildingBlocksSidebar/index.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { useState } from "react";
 
-import { DEFAULT_AGENT_CONTEXT } from "./agentChat";
 import type { BuildingBlock, BuildingBlockCategory } from "./index";
 import { BuildingBlocksSidebar } from "./index";
 import React from "react";
@@ -85,9 +84,6 @@ const sampleBlocks: BuildingBlockCategory[] = [
 const meta = {
   title: "ui/BuildingBlocksSidebar",
   component: BuildingBlocksSidebar,
-  args: {
-    agentContext: DEFAULT_AGENT_CONTEXT,
-  },
   parameters: {
     layout: "fullscreen",
   },
